backend/krupal_chatbot/requirements.txt
streamlit
google-generativeai
langchain
langchain-community
langchain-ollama
faiss-cpu
python-dotenv
